refactor(api): move authenticateForSaml controller in IAM context
